feat: move UserExternalIdentitiesViewModel to commons (batch 4)
